+++ src/product/tool-definitions.js
@@ -44,6 +44,9 @@ import { audioFilterToolDefinitions } from './definitions/audio-filter-tools.js'
 import { projectManagementDefinitions } from './definitions/project-management.js';
 import { imageFileToolDefinitions } from './definitions/image-file-tools.js';
 import { imageExtraToolDefinitions } from './definitions/image-extra-tools.js';
+import { imageExifToolDefinitions } from './definitions/image-exif-tools.js';
+import { imageAnalysisExtraToolDefinitions } from './definitions/image-analysis-extra-tools.js';
+import { imageLayoutToolDefinitions } from './definitions/image-layout-tools.js';
 import { imageEditingToolDefinitions } from './definitions/image-editing-tools.js';
 import { imageEnhancementToolDefinitions } from './definitions/image-enhancement-tools.js';
 import { imageFilterToolDefinitions } from './definitions/image-filter-tools.js';
@@ -1090,6 +1093,9 @@ const allToolDefinitions = Object.freeze({
     ...projectManagementDefinitions,
     ...imageFileToolDefinitions,
     ...imageExtraToolDefinitions,
+    ...imageExifToolDefinitions,
+    ...imageAnalysisExtraToolDefinitions,
+    ...imageLayoutToolDefinitions,
     ...imageEditingToolDefinitions,
     ...imageEnhancementToolDefinitions,
     ...imageFilterToolDefinitions,